PWM+DIR signal tester:
A test firmware that you can use to verify if the hardware is working or not. Please find attached file and load it to SimuCUBE with DfuSe as instructed here.

The firmware will do following:
-blink 3 LEDs of SimuCUBE (2 sec on, 2 sec off and repeat)
-toggle status of HSIN1 and HSIN2 of IONI (these are the PWM+DIR signals) at same 2 sec on/off pattern

To see if hardware works, connect with Granity and check if the HSIN inputs are toggling.